Pulled the snippet search utility methods up into the common test base class and updated the folding range tests to use them as well.
